src/requirements.txt
colorama==0.4.5
psycopg2-binary==2.9.4
py-mon==1.1.0
watchdog==2.1.9
pandas